# src/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CreateOrderItemRequest:
    properties:
      book_id:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
      - book_id
      - quantity
    type: object
  dtos.CreateOrderRequest:
    properties:
      order_items:
        items:
          $ref: '#/definitions/dtos.CreateOrderItemRequest'
        type: array
      payment_id:
        type: string
      promotion_id:
        type: string
      shipping_address:
        type: string
    required:
      - order_items
      - payment_id
      - shipping_address
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      error: {}
      headers: {}
      message:
        type: string
      pagination: {}
      status_code:
        type: integer
    type: object
host: localhost:5000
info:
  contact:
    email: tntuan0910@gmail.com
    name: tuan.tran
  description: This is the API documentation for the Rainbow API.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Rainbow API
  version: '1.0'
paths:
...
  /orders:
    get:
      consumes:
        - application/json
      description: Get orders by user ID
      parameters:
        - default: 1
          description: Page number
          in: query
          name: page
          type: integer
        - default: 10
          description: Number of items per page
          in: query
          name: limit
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get orders by user ID
      tags:
        - orders
    post:
      consumes:
        - application/json
      description: Create order
      parameters:
        - description: Order request
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/dtos.CreateOrderRequest'
      produces:
        - application/json
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Create order
      tags:
        - orders
  /orders/{id}:
    delete:
      consumes:
        - application/json
      description: Delete order by its ID
      parameters:
        - description: Order ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        '204':
          description: No Content
          schema:
            $ref: '#/definitions/response.APIResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Delete order
      tags:
        - orders
    get:
      consumes:
        - application/json
      description: Get order by its ID
      parameters:
        - description: Order ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get an order
      tags:
        - orders
...
swagger: '2.0'